15_GHOST_LAG/requirements.txt
openai==1.3.0
anthropic==0.7.0
transformers==4.30.2
nltk==3.8.1
spacy==3.6.0
textblob==0.17.1
gensim==4.3.1
scikit-learn==1.3.2
pandas==2.0.3
numpy==1.24.3
aiohttp==3.8.5
python-dotenv==1.0.0
fastapi==0.100.0
uvicorn==0.23.1
python-multipart==0.0.6
pydantic==2.0.3
SQLAlchemy==2.0.19
aiosqlite==0.19.0
markdown==3.4.3
beautifulsoup4==4.12.2
python-docx==0.8.11
pypdf2==3.0.1
ebooklib==0.18
pytz==2023.3
schedule==1.2.0
tenacity==8.2.2
python-dateutil==2.8.2
watchdog==3.0.0
gitpython==3.1.32
pillow==10.0.0
reportlab==4.0.4
fpdf2==2.7.4
mammoth==1.6.0
regex==2023.6.3
language-tool-python==2.7.1
py-readability-metrics==1.4.5 